Scripting API: rename underlined to underline
Although `underlined` is grammatically correct, `underline` is used to align with naming conventions in the Groovy Scripting API for other formatting settings. For example, the `bold` property uses method names like `isBold()`, `setBold(isBold)``, and `isBoldSet()``, which integrate naturally into Groovy code such as `node.style.font.bold = true`. This is in contrast with `bolded` or `emboldened`, which are the corresponding adjectives for `bold`.

Formatting to adjective
-----------------------
- Bold: bolded, emboldened
- Italic: italicized
- Underline: underlined
- Strikethrough: struck-through, crossed-out